+++ server/controllers/orderController.js
@@ -13,7 +13,9 @@ products.price
 FROM cart
 JOIN products
 ON cart.product_id = products.id
-`);
+WHERE cart.user_id = $1
+`,
+[req.user.id]);
 
 if (cart.rows.length === 0) {
   return res.status(400).json({
@@ -33,10 +35,10 @@ if (cart.rows.length === 0) {
     // Create Order
 
     const order = await pool.query(
-      `INSERT INTO orders (total_amount)
-      VALUES ($1)
+      `INSERT INTO orders (user_id, total_amount)
+      VALUES ($1, $2)
       RETURNING *`,
-      [totalAmount]
+      [req.user.id, totalAmount,]
     );
 
     // Step 4
@@ -58,7 +60,9 @@ if (cart.rows.length === 0) {
 
     // Step 5
     // Clear Cart
-     await pool.query("DELETE FROM cart");
+     await pool.query("DELETE FROM cart WHERE user_id = $1",
+      [req.user.id]
+     );
 
     // Step 6
     // Send Response
@@ -78,6 +82,166 @@ if (cart.rows.length === 0) {
   }
 };
 
+const getOrders = async (req, res) => {
+  try {
+
+    let result;
+
+    if (req.user.role === "admin") {
+
+      result = await pool.query(`
+        SELECT *
+        FROM orders
+        ORDER BY created_at DESC
+      `);
+
+    } else {
+
+      result = await pool.query(
+        `
+        SELECT *
+        FROM orders
+        WHERE user_id = $1
+        ORDER BY created_at DESC
+        `,
+        [req.user.id]
+      );
+
+    }
+
+    res.json(result.rows);
+
+  } catch (error) {
+
+    console.error(error);
+
+    res.status(500).json({
+      message: "Server Error",
+    });
+
+  }
+};
+
+const getSingleOrder = async (req, res) => {
+  try {
+
+    const { id } = req.params;
+
+    let order;
+
+    if (req.user.role === "admin") {
+
+      order = await pool.query(
+        `
+        SELECT *
+        FROM orders
+        WHERE id = $1
+        `,
+        [id]
+      );
+
+    } else {
+
+      order = await pool.query(
+        `
+        SELECT *
+        FROM orders
+        WHERE id = $1
+        AND user_id = $2
+        `,
+        [id, req.user.id]
+      );
+
+    }
+
+    if (order.rows.length === 0) {
+      return res.status(404).json({
+        message: "Order not found",
+      });
+    }
+
+    const items = await pool.query(
+      `
+      SELECT
+      order_items.id,
+      order_items.quantity,
+      order_items.price,
+      products.title,
+      products.image
+      FROM order_items
+      JOIN products
+      ON order_items.product_id = products.id
+      WHERE order_items.order_id = $1
+      `,
+      [id]
+    );
+
+    res.json({
+      order: order.rows[0],
+      items: items.rows,
+    });
+
+  } catch (error) {
+    console.error(error);
+
+    res.status(500).json({
+      message: "Server Error",
+    });
+  }
+};
+
+const updateOrderStatus = async (req, res) => {
+
+  try {
+
+    const { id } = req.params;
+
+    const { status } = req.body;
+
+
+    const result = await pool.query(
+
+      `
+      UPDATE orders
+
+      SET status = $1
+
+      WHERE id = $2
+
+      RETURNING *
+      `,
+
+      [status, id]
+
+    );
+
+
+    if (result.rows.length === 0) {
+
+      return res.status(404).json({
+        message: "Order not found",
+      });
+
+    }
+
+
+    res.json(result.rows[0]);
+
+  } catch (error) {
+
+    console.error(error);
+
+    res.status(500).json({
+      message: "Server Error",
+    });
+
+  }
+
+};
+
 module.exports = {
   checkout,
+  getOrders,
+  getSingleOrder,
+  updateOrderStatus,
 };
